+++ src/store/reducer.js
@@ -2,7 +2,7 @@ import {combineReducers} from "redux";
 import {reducer as offers} from "./offers/offers.js";
 import {reducer as cities} from "./cities/reducer.js";
 import {reducer as reviews} from "./reviews/reviews.js";
-import {reducer as nearby} from "./nearby/nearby.js";
+import {reducer as nearby} from "./nearby/reducer.js";
 import NameSpace from "./name-space.js";
 
 export default combineReducers({
